ACES 1.2 Colourimetric Display Edition
This is an untouched version of the ACES
1.2 generated version at Thomas Mansencal's
repository at:
https://github.com/colour-science/OpenColorIO-Configs

The net sum transform chain is 100% intact,
and this repository only uses some of the
configuration options available via the Python
generation to make the configuration more sane
for people using software that properly abides
by OpenColorIO's design intention of discrete
colourimetric displays.

The TL;DR version is that this edition does
not create a gabillion sloppy fucking transforms
in one gargantuan mess.

The reason the existing ACES configuration
defaults to the ridiculously sloppy mess is
largely due to applications not supporting
OpenColorIO correctly (hello Nuke), so the
lowest common denominator is targeted, and
you, the end audience member, ends up with
that aforementionned sloppy fucking mess.

To the folks who read this around ACES, you
know who you are, do the right thing and force
the sloppy fucking applications to get their
shit together and fix their broken ass software.

Lead by example.

The command used to generate this configuration
is as follows:

aces_1.2/python/bin/create_aces_config -a
aces-dev/transforms/ctl --lutResolution1d 4096
--lutResolution3d 65 -c aces_1.2
--createMultipleDisplays --dontBakeSecondaryLUTs